+++ src/OneBlinkFormBase.tsx
@@ -560,7 +560,14 @@ function OneBlinkFormBase({
 
   const handleChange = React.useCallback<FormElementValueChangeHandler>(
     (element, value) => {
-      if (disabled || element.type === 'page' || element.type === 'section') {
+      if (
+        //This will ensure on a read only form that the summary element can still be
+        //displayed as it needs handleChange so it can render due to the dynamic
+        //nature of the summary element.
+        (disabled && element.type !== 'summary') ||
+        element.type === 'page' ||
+        element.type === 'section'
+      ) {
         return
       }